Introduced more parameters for find_neighbors
In the past state there was only line_margin
which was used to determine vertical, horizontal
and height distance checks. Now these values are
separated. This brings flexibility for the
calculation which lines are part of a box.

The old line_margin is used for vertical line
distance. The check for same height was disabled
because it led to multiple nested text boxes.
Finding a reasonable value that suits all PDFs is
quit difficult.